And update vendor

43ee1abaca PHPSessionHandler: Suppress warnings in initialize()
568bd6d1bf Treat all time values on Special:Watchlist as floats
b9cc61f6c4 Guard against uncountable tag values
0d57b370df Fix total breakage of SQLite web upgrade
633252c3e9 Use LB server configuration to force DB domains in ExternalStorageDB
c6f0a3804d Make ExternalStoreDB "wiki" context override the server "dbname" field
4bd19c0486 Fix pingback on non-MySQL
5e8e72f822 UsersPager: Fix typo in formDescriptor array
44686573c9 SECURITY: Make 'newbie' limit in $wgRateLimits really override 'user' limit
6f98988380 SECURITY: Do not allow botpassword login if account locked.
845dee9fa4 SECURITY: Don't exclude .htaccess files from `git archive`
a4c8065634 Bump 1.31.1
1f664ea4eb WatchedItemStore::countVisitingWatchersMultiple() shouldn't query all titles when asked for none
8641df974d Document removal of CologneBlue & Modern from the tarball
590e94d474 Fix typo in postgres patch-drop-ar_text.sql
a9c02b13f6 Add pear/Net_SMTP 1.7.3 to composer dependencies
8f6ecc4df8 Load installer i18n when running update.php
70532ff53f Remove deprecated pear/mail_mime-decode from composer dependancies
5cab354194 Update git submodules
942728ab20 Use "break" instead of "continue"
935402afae Update git submodules
8635864ed0 Use "break" instead of "continue" inside a switch
18fcb2bf29 LocalisationCache: Avoid use of compact()
f80e50f318 Fix PHP warnings "preg_replace(): [...] invalid range in character class"
14d08d1df9 Avoid PHP 7.2 warnings in DBConRefTest about count() on non-Countable
c7246fc201 Suppress "Headers already sent" in PHP 7.2 too
7a547df7c0 Output only to stderr in unit tests
c1c0d24352 Add session_write_close() calls to SessionManager tests
048089ae3a Update ImportableUploadRevisionImporter for interwiki usernames
e62a8361d7 Include IP address in "Login for $1 succeeded" log entry
6bc0e36d8b Don't pass a MailAddress pass the email to mail()
886f3307da installer: Don't link to the obsolete "Extension Matrix" page
7065200b03 Database: Allow selectFieldValues() to accept SQL fragments
90232b6f36 SECURITY: Disallow loading JS/CSS/Json subpages from unregistered users and log
a5e349c908 SECURITY: Don't allow loading unprotected JS files
b1baa56b84 <ins>/<del> elements can be phrasing or flow
25b0bce762 RemexCompatMunger: Don't call endTag() in case B/b
c26c3ba1a5 Upgrade wikimedia/remex-html to 2.0.0
2c65b58fa7 Upgrade wikimedia/remex-html to 2.0.1
4293f8fa94 Update git submodules
145d94f9bd Update git submodules
a9f342c0b4 Update wikimedia/base-convert to 2.0.0
5b69f06a4b Update git submodules
23bf104160 Don't throw E_NOTICEs about undefined properties
921e26598f composer.json, require ext-fileinfo
a013eed7d8 Update git submodules
29ef5f88ab Completion search should not change the search query
b0b7727e5e Add test for completionSearch with wgCapitalLinkOverrides
107ee818cf Update git submodules
8e4e0d8922 Update git submodules
4b70558f4a Use $revQuery['joins'] in query in populateSearchIndex
f74dcc34a1 SQL syntax error in MS-SQL file
7a6393fc2b LogFormatter: Fail softer when trying to link an invalid titles
63f8c9aad4 RELEASE-NOTES-1.31: Add in other cherry-picks since 1.31.1 was cut
110f9e3db7 Save value from CLI installers `--lang` argument
141cb48654 ImageListPager: Actor migration for buildQueryConds()
021ed39d59 Use our fork of less.php
c7cd5e512c Upgrade wikimedia/ip-set to 1.3.0
d88136612b Update git submodules
be5735b694 Update git submodules
cbc99caf67 i18n: Clarify the default sidebar 'Help' link is about MediaWiki itself
5a4eb3dd50 Fix copy-paste error
2958311c4f Update git submodules
8caf853f22 Add join conditions to ActiveUsersPager
836f10dcbd User: Bypass repeatable-read when creating an actor_id

+++ includes/specials/SpecialWatchlist.php
@@ -772,7 +772,7 @@ class SpecialWatchlist extends ChangesListSpecialPage {
 		] ) );
 		asort( $hours );
 
-		$select = new XmlSelect( 'days', 'days', $selectedHours / 24 );
+		$select = new XmlSelect( 'days', 'days', (float)( $selectedHours / 24 ) );
 
 		foreach ( $hours as $value ) {
 			if ( $value < 24 ) {
@@ -780,7 +780,7 @@ class SpecialWatchlist extends ChangesListSpecialPage {
 			} else {
 				$name = $this->msg( 'days' )->numParams( $value / 24 )->text();
 			}
-			$select->addOption( $name, $value / 24 );
+			$select->addOption( $name, (float)( $value / 24 ) );
 		}
 
 		return $select->getHTML() . "\n<br />\n";
